feat: add model

+++ internal/model/group_info.go
@@ -17,7 +17,7 @@ type GroupInfo struct {
 	Members     json.RawMessage `gorm:"column:members"`                // 用户uuid
 	MemberCount int             `gorm:"column:member_count;default:1"` // 群聊人数【默认群主1人】
 	OwnerID     string          `gorm:"column:owner_id"`               // 群主uuid
-	AddMode     int             `gorm:"column:add_mode"`               // 0 直接 1 审核
+	AddMode     int8            `gorm:"column:add_mode"`               // 0 直接 1 审核
 	Status      int8            `gorm:"column:status"`                 // 0 正常 1 禁用 2 解散
 	CreatedAt   time.Time
 	DeletedAt   gorm.DeletedAt `gorm:"index"`
